requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.8
mysqlclient==1.4.6
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
sightengine==1.4.0
sqlparse==0.3.0
urllib3==1.25.8
